# feegow-challenge-covid-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Vacinação de Funcionários",
        "description": "Uma API para registro de vacinação de funcionários contra a COVID-19.",
        "contact": {
            "name": "Suporte Técnico",
            "email": "suporte@empresa.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/funcionario-vacina": {
            "get": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Lista todas as vacinações de funcionários",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Lista de vacinações de funcionários"
                    }
                }
            },
            "post": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Registra uma nova vacinação para um funcionário",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "funcionario_cpf",
                                    "vacina_id",
                                    "data_dose",
                                    "dose"
                                ],
                                "properties": {
                                    "funcionario_cpf": {
                                        "type": "string",
                                        "example": "12345678900"
                                    },
                                    "vacina_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "data_dose": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-14"
                                    },
                                    "dose": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vacinação registrada com sucesso!"
                    }
                }
            }
        },
        "/funcionario-vacina/{funcionario_id}/{vacina_id}": {
            "get": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Retorna detalhes de uma vacinação específica de um funcionário",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "funcionario_id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da vacinação"
                    },
                    "404": {
                        "description": "Vacinação não encontrada"
                    }
                }
            }
        },
        "/funcionario-vacina/vacina/{vacina_id}": {
            "get": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Lista todas as vacinações de funcionários por ID de vacina",
                "operationId": "indexByVacinaId",
                "parameters": [
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vacinações de funcionários por ID de vacina"
                    },
                    "404": {
                        "description": "Nenhuma vacinação de funcionário encontrada"
                    }
                }
            }
        },
        "/funcionario-vacina/funcionario/{funcionario_id}": {
            "get": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Lista todas as vacinações de um funcionário por ID de funcionário",
                "operationId": "indexByFuncionarioId",
                "parameters": [
                    {
                        "name": "funcionario_id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vacinações de um funcionário por ID de funcionário"
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                }
            }
        },
        "/funcionario-vacina/funcionario/cpf/{funcionario_cpf}": {
            "get": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Lista todas as vacinações de um funcionário por CPF de funcionário",
                "operationId": "indexByFuncionarioCpf",
                "parameters": [
                    {
                        "name": "funcionario_cpf",
                        "in": "path",
                        "description": "CPF do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vacinações de um funcionário por CPF de funcionário"
                    },
                    "404": {
                        "description": "Funcionário não encontrado"
                    }
                }
            }
        },
        "/funcionario-vacina/{funcionario_cpf}/{vacina_id}/{dose}": {
            "put": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Atualiza os detalhes de uma vacinação específica de um funcionário",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "funcionario_cpf",
                        "in": "path",
                        "description": "CPF do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dose",
                        "in": "path",
                        "description": "Dose da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data_dose"
                                ],
                                "properties": {
                                    "data_dose": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-14"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro de vacinação atualizado com sucesso!"
                    }
                }
            },
            "delete": {
                "tags": [
                    "FuncionarioVacina"
                ],
                "summary": "Remove uma vacinação específica de um funcionário",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "funcionario_cpf",
                        "in": "path",
                        "description": "CPF do funcionário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vacina_id",
                        "in": "path",
                        "description": "ID da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dose",
                        "in": "path",
                        "description": "Dose da vacina",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Registro de vacinação removido com sucesso!"
                    }
                }
            }
        },
...
    }
}